possibly fix debian error on time_t
I am a bit concerned that this may break gri on other platforms, but not
really very concerned, since this fiddling with time types is likely
irrelevant to modern machines.  And I doubt that anybody is building gri
on some of the platforms for which the code is making special cases.

Please note that I do not have access to a machine that reproduces the
error message.  I am on macos, and don't want to break my toolchain by
over-writing my C++ compilers, etc., to try to reproduce the bug.

Lacking access to any type of linux machine, I am not able to test
whether my change fixes the problem. But the `git diff` is only one line
long, so hopefully if there is still a problem on debian, someone can
tell me a preprocesser symbol that I can use (and perhaps a header.h
file I can specify) to define the `time_t` type.

Sorry for the hassles.

Dan Kelley